revisao 2

+++ bases/br_inpe_prodes/desmatamento_municipio/publish.sql
@@ -23,12 +23,12 @@ CREATE VIEW basedosdados-dev.br_inpe_prodes.desmatamento_municipio AS
 SELECT 
 SAFE_CAST(ano AS INT64) ano,
 SAFE_CAST(id_municipio AS STRING) id_municipio,
-SAFE_CAST(area AS NUMERIC) area,
-SAFE_CAST(desmatado AS NUMERIC) desmatado,
-SAFE_CAST(incremento AS NUMERIC) incremento,
-SAFE_CAST(floresta AS NUMERIC) floresta,
-SAFE_CAST(nuvem AS NUMERIC) nuvem,
-SAFE_CAST(nao_observado AS NUMERIC) nao_observado,
-SAFE_CAST(nao_floresta AS NUMERIC) nao_floresta,
-SAFE_CAST(hidrografia AS NUMERIC) hidrografia
+SAFE_CAST(area AS FLOAT64) area,
+SAFE_CAST(desmatado AS FLOAT64) desmatado,
+SAFE_CAST(incremento AS FLOAT64) incremento,
+SAFE_CAST(floresta AS FLOAT64) floresta,
+SAFE_CAST(nuvem AS FLOAT64) nuvem,
+SAFE_CAST(nao_observado AS FLOAT64) nao_observado,
+SAFE_CAST(nao_floresta AS FLOAT64) nao_floresta,
+SAFE_CAST(hidrografia AS FLOAT64) hidrografia
 from basedosdados-dev.br_inpe_prodes_staging.desmatamento_municipio as t
